+++ src/dakara_feeder/__init__.py
@@ -1,4 +1,6 @@
-from dakara_feeder.dakara_feeder import DakaraFeeder
+from dakara_feeder.songs_feeder import SongsFeeder
+from dakara_feeder.tags_feeder import TagsFeeder
 from dakara_feeder.version import __date__, __version__
+from dakara_feeder.work_types_feeder import WorkTypesFeeder
 
-__all__ = ["DakaraFeeder", "__version__", "__date__"]
+__all__ = ["SongsFeeder", "TagsFeeder", "WorkTypesFeeder", "__version__", "__date__"]