+++ src/dakara_feeder/dakara_server.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from dakara_base.http_client import HTTPClient
+from dakara_base.http_client import HTTPClient, ResponseInvalidError
 from path import Path
 
 logger = logging.getLogger(__name__)
@@ -12,12 +12,12 @@ class DakaraServer(HTTPClient):
     """Client to the Dakara server."""
 
     def get_songs(self):
-        """Retreive the songs of the library containing their path
+        """Retreive the songs of the library containing their path.
 
         Returns:
             list: List of path on the songs.
         """
-        endpoint = "library/feeder/retrieve/"
+        endpoint = "library/songs/retrieve/"
         songs = self.get(endpoint)
 
         # join the directory and the filename
@@ -53,3 +53,81 @@ class DakaraServer(HTTPClient):
         """
         endpoint = "library/songs/{}/".format(song_id)
         self.put(endpoint, json=song)
+
+    def prune_artists(self):
+        """Prune artists without songs.
+
+        Returns:
+            int: Number of deleted artists.
+        """
+        endpoint = "library/artists/prune/"
+        return self.delete(endpoint)["deleted_count"]
+
+    def prune_works(self):
+        """Prune works without songs.
+
+        Return:
+            int: Number of deleted works.
+        """
+        endpoint = "library/works/prune/"
+        return self.delete(endpoint)["deleted_count"]
+
+    def create_tag(self, tag):
+        """Create a tag on the server.
+
+        Args:
+            tag (dict): JSON representation of a tag.
+
+        Raises:
+            TagAlreadyExistsError: If the tag exists on the server, i.e. if the
+                server returns 400.
+            dakara_base.http_client.ResponseInvalidError: If the response of
+                the server is not OK.
+        """
+
+        def on_error(response):
+            if response.status_code == 400:
+                return TagAlreadyExistsError()
+
+            return ResponseInvalidError(
+                "Error {} when communicating with the server: {}".format(
+                    response.status_code, response.text
+                )
+            )
+
+        endpoint = "library/song-tags/"
+        self.post(endpoint, tag, function_on_error=on_error)
+
+    def create_work_type(self, work_type):
+        """Create a work type on the server.
+
+        Args:
+            work_type (dict): JSON representation of a work type.
+
+        Raises:
+            WorkTypeAlreadyExistsError: If the work type exists on the server,
+                i.e. if the server returns 400.
+            dakara_base.http_client.ResponseInvalidError: If the response of
+                the server is not OK.
+        """
+
+        def on_error(response):
+            if response.status_code == 400:
+                return WorkTypeAlreadyExistsError()
+
+            return ResponseInvalidError(
+                "Error {} when communicating with the server: {}".format(
+                    response.status_code, response.text
+                )
+            )
+
+        endpoint = "library/work-types/"
+        self.post(endpoint, work_type, function_on_error=on_error)
+
+
+class TagAlreadyExistsError(Exception):
+    """Error if a tag already exists."""
+
+
+class WorkTypeAlreadyExistsError(Exception):
+    """Error if a work type already exists."""