+++ tests/unit/test_dakara_server.py
@@ -1,5 +1,5 @@
 from unittest import TestCase
-from unittest.mock import patch
+from unittest.mock import ANY, patch
 
 from path import Path
 
@@ -57,7 +57,7 @@ class DakaraServerTestCase(TestCase):
         )
 
         # assert the mock
-        mocked_get.assert_called_with("library/feeder/retrieve/")
+        mocked_get.assert_called_with("library/songs/retrieve/")
 
     @patch.object(dakara_server.DakaraServer, "post", autoset=True)
     def test_post_song(self, mocked_post):
@@ -122,3 +122,188 @@ class DakaraServerTestCase(TestCase):
 
         # assert the mock
         mocked_put.assert_called_with("library/songs/42/", json=song)
+
+    @patch.object(dakara_server.DakaraServer, "delete", autoset=True)
+    def test_prune_artists(self, mocked_delete):
+        """Test to prune artists."""
+        # mock objects
+        mocked_delete.return_value = {"deleted_count": 2}
+
+        # create the object
+        server = dakara_server.DakaraServer(
+            self.config, endpoint_prefix=self.endpoint_prefix
+        )
+
+        # call the method
+        deleted_count = server.prune_artists()
+
+        # assert the value
+        self.assertEqual(deleted_count, 2)
+
+        # assert the mock
+        mocked_delete.assert_called_with("library/artists/prune/")
+
+    @patch.object(dakara_server.DakaraServer, "delete", autoset=True)
+    def test_prune_works(self, mocked_delete):
+        """Test to prune works."""
+        # mock objects
+        mocked_delete.return_value = {"deleted_count": 2}
+
+        # create the object
+        server = dakara_server.DakaraServer(
+            self.config, endpoint_prefix=self.endpoint_prefix
+        )
+
+        # call the method
+        deleted_count = server.prune_works()
+
+        # assert the value
+        self.assertEqual(deleted_count, 2)
+
+        # assert the mock
+        mocked_delete.assert_called_with("library/works/prune/")
+
+    @patch.object(dakara_server.DakaraServer, "post", autoset=True)
+    def test_create_tag(self, mocked_post):
+        """Test to create tag."""
+        # create the object
+        server = dakara_server.DakaraServer(
+            self.config, endpoint_prefix=self.endpoint_prefix
+        )
+
+        # create tag
+        tag = {"name": "tag1", "color_hue": 250}
+
+        # call the method
+        server.create_tag(tag)
+
+        # assert the call
+        mocked_post.assert_called_with("library/song-tags/", tag, function_on_error=ANY)
+
+    @patch("dakara_base.http_client.requests.post", autoset=True)
+    def test_create_tag_error_already_exists(self, mocked_post):
+        """Test to create tag that already exists."""
+        # create the mock
+        mocked_post.return_value.ok = False
+        mocked_post.return_value.status_code = 400
+
+        # create the object
+        server = dakara_server.DakaraServer(
+            self.config, endpoint_prefix=self.endpoint_prefix
+        )
+
+        # artificially connect the server
+        server.token = "token"
+
+        # create tag
+        tag = {"name": "tag1", "color_hue": 250}
+
+        # call the method
+        with self.assertRaises(dakara_server.TagAlreadyExistsError):
+            server.create_tag(tag)
+
+        # assert the call
+        mocked_post.assert_called_with(
+            self.url + "library/song-tags/", tag, headers=ANY
+        )
+
+    @patch("dakara_base.http_client.requests.post", autoset=True)
+    def test_create_tag_error_other(self, mocked_post):
+        """Test an unknown problem when creating a tag."""
+        # create the mock
+        mocked_post.return_value.ok = False
+        mocked_post.return_value.status_code = 999
+        mocked_post.return_value.text = "error message"
+
+        # create the object
+        server = dakara_server.DakaraServer(
+            self.config, endpoint_prefix=self.endpoint_prefix
+        )
+
+        # artificially connect the server
+        server.token = "token"
+
+        # create tag
+        tag = {"name": "tag1", "color_hue": 250}
+
+        # call the method
+        with self.assertRaises(dakara_server.ResponseInvalidError) as error:
+            server.create_tag(tag)
+
+        # assert the error
+        self.assertEqual(
+            str(error.exception),
+            "Error 999 when communicating with the server: error message",
+        )
+
+    @patch.object(dakara_server.DakaraServer, "post", autoset=True)
+    def test_create_work_type(self, mocked_post):
+        """Test to create work type."""
+        # create the object
+        server = dakara_server.DakaraServer(
+            self.config, endpoint_prefix=self.endpoint_prefix
+        )
+
+        # create work type
+        work_type = {"query_name": "wt1", "name": "Work Type 1"}
+
+        # call the method
+        server.create_work_type(work_type)
+
+        # assert the call
+        mocked_post.assert_called_with(
+            "library/work-types/", work_type, function_on_error=ANY
+        )
+
+    @patch("dakara_base.http_client.requests.post", autoset=True)
+    def test_create_work_type_error_already_exists(self, mocked_post):
+        """Test to create work type that already exists."""
+        # create the mock
+        mocked_post.return_value.ok = False
+        mocked_post.return_value.status_code = 400
+
+        # create the object
+        server = dakara_server.DakaraServer(
+            self.config, endpoint_prefix=self.endpoint_prefix
+        )
+
+        # artificially connect the server
+        server.token = "token"
+
+        # create work type
+        work_type = {"query_name": "wt1", "name": "Work Type 1"}
+
+        # call the method
+        with self.assertRaises(dakara_server.WorkTypeAlreadyExistsError):
+            server.create_work_type(work_type)
+
+        # assert the call
+        mocked_post.assert_called_with(
+            self.url + "library/work-types/", work_type, headers=ANY
+        )
+
+    @patch("dakara_base.http_client.requests.post", autoset=True)
+    def test_create_work_type_error_other(self, mocked_post):
+        """Test an unknown problem when creating a work type."""
+        # create the mock
+        mocked_post.return_value.ok = False
+        mocked_post.return_value.status_code = 999
+        mocked_post.return_value.text = "error message"
+
+        # create the object
+        server = dakara_server.DakaraServer(
+            self.config, endpoint_prefix=self.endpoint_prefix
+        )
+
+        # artificially connect the server
+        server.token = "token"
+
+        # create work type
+        work_type = {"query_name": "wt1", "name": "Work Type 1"}
+
+        # call the method
+        with self.assertRaisesRegex(
+            dakara_server.ResponseInvalidError,
+            "Error 999 when communicating with the server: error message",
+        ):
+            server.create_work_type(work_type)